EES-7023 Tidy up Content API's 'Publication infos' endpoint (#6920)
* EES-7023 Add comments to the ListPublicationInfos endpoint and response types PublicationInfoViewModel and ReleaseInfoViewModel to make it clear that they are only used by the Search Docs Function App.

* EES-7023 Add required keyword to PublicationId and LatestPublishedRelease properties of PublicationInfoViewModel

* EES-7023 Make LatestPublishedRelease property type in PublicationInfoViewModel non-optional and refactor ReleaseInfoViewModel to be PublicationInfoReleaseViewModel.

* EES-7023 Tidy up ListPublicationInfos in PublicationController.
