--- src/GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp/Clients/ContentApi/Dtos/ReleaseInfoDto.cs
@@ -1,7 +1,0 @@
-namespace GovUk.Education.ExploreEducationStatistics.Content.Search.FunctionApp.Clients.ContentApi.Dtos;
-
-public record ReleaseInfoDto
-{
-    public Guid? ReleaseId { get; init; }
-    public string? ReleaseSlug { get; init; }
-}
