--- src/GovUk.Education.ExploreEducationStatistics.Content.ViewModels/ReleaseInfoViewModel.cs
@@ -1,7 +1,0 @@
-﻿namespace GovUk.Education.ExploreEducationStatistics.Content.ViewModels;
-
-public record ReleaseInfoViewModel
-{
-    public required Guid ReleaseId { get; init; }
-    public required string ReleaseSlug { get; init; }
-}
